requirements.txt
# python 3.11.5

aiofiles==23.2.1
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.5
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
nest-asyncio==1.5.7
orjson==3.9.7
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
twitter-api-client==0.10.11
uvloop==0.17.0
websockets==11.0.3
django-rest-framework==0.1.0
